Refactor(activity): 활동 페이지 관련 컴포넌트 폴더 구조 수정
