+++ src/app/(main)/activity/[semesterName]/[activityId]/_components/activity/index.ts
@@ -1,0 +1,1 @@
+export { ActivityList } from './ActivityList'
