+++ src/app/(main)/activity/[semesterName]/[activityId]/_components/board-list/index.ts
@@ -1,0 +1,1 @@
+export { ActivityBoardList } from './BoardList'
