+++ src/app/(main)/activity/[semesterName]/[activityId]/_components/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { default as ActivityErrorFallback } from './ErrorFallback'
