+++ src/app/(main)/activity/[semesterName]/[activityId]/_components/index.ts
@@ -1,0 +1,3 @@
+export * from './activity'
+export * from './error-fallback'
+export * from './board-list'
