+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/BoardHero/index.tsx
@@ -4,6 +4,7 @@ import { useQuery } from '@tanstack/react-query'
 import { useAction } from 'next-safe-action/hooks'
 import { usePathname, useRouter } from 'next/navigation'
 
+import { ActivityBreadcrumb } from '@/components/feature'
 import { Button } from '@/components/ui/button'
 import { Separator } from '@/components/ui/separator'
 import { useToast } from '@/components/ui/use-toast'
@@ -12,8 +13,6 @@ import { queryClient } from '@/service/components/ReactQueryClientProvider'
 import { boardDetailQuery } from '@/service/data/boards'
 import { deleteBoardAction } from '@/service/server/board/delete-board'
 import { useMyInfoStore } from '@/store/myInfo'
-
-import { ActivityBreadcrumb } from '~activity/_components/ActivityBreadcrumb'
 
 import { BoardDetail } from './BoardDetail'
 import { BoardHeroSkeleton } from './BoardHeroSkeleton'
