+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/posts/[postId]/_components/ActivityPostHero/index.tsx
@@ -3,11 +3,10 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { usePathname } from 'next/navigation'
 
+import { ActivityBreadcrumb } from '@/components/feature'
 import { Separator } from '@/components/ui/separator'
 import { DATA_ERROR_MESSAGES } from '@/constant/errorMessage'
 import { boardDetailQuery } from '@/service/data/boards'
-
-import { ActivityBreadcrumb } from '~activity/_components/ActivityBreadcrumb'
 
 type ActivityPostHeroProps = {
   activityId: number
