+++ src/app/(main)/activity/[semesterName]/_components/index.ts
@@ -1,0 +1,1 @@
+export { RedirectActivity } from './RedirectActivity'
