+++ src/app/(main)/admin/semester/_components/index.ts
@@ -1,3 +1,2 @@
 export * from './activity-accordion'
 export * from './semester-list'
-export * from './semester-skeleton'
