--- src/app/(main)/admin/semester/_components/semester-skeleton/SemesterSkeleton.tsx
@@ -1,9 +1,0 @@
-export const SemesterSkeleton = () => {
-  return (
-    <div className="flex gap-3">
-      <div className="h-9 w-24 rounded-full bg-slate-100"></div>
-      <div className="h-9 w-24 rounded-full bg-slate-50"></div>
-      <div className="h-9 w-24 rounded-full bg-slate-200/70"></div>
-    </div>
-  )
-}
