--- src/app/(main)/admin/semester/_components/semester-skeleton/index.ts
@@ -1,1 +1,0 @@
-export { SemesterSkeleton } from './SemesterSkeleton'
