+++ src/app/(main)/admin/semester/page.tsx
@@ -1,13 +1,9 @@
 'use client'
 
-import { SectionWithTitle } from '@/components/feature'
+import { AdminSemesterSkeleton, SectionWithTitle } from '@/components/feature'
 import { useGetSemesters } from '@/service/data/semester'
 
-import {
-  ActivityAccordion,
-  SemesterList,
-  SemesterSkeleton,
-} from './_components'
+import { ActivityAccordion, SemesterList } from './_components'
 
 const AdminSemesterPage = () => {
   const { semesters, status } = useGetSemesters()
@@ -32,7 +28,7 @@ const AdminSemesterPageSkeleton = () => {
   return (
     <div className="flex w-full flex-col items-center">
       <SectionWithTitle title="학기 관리">
-        <SemesterSkeleton />
+        <AdminSemesterSkeleton />
       </SectionWithTitle>
       <SectionWithTitle title="활동 관리">
         <div />
