+++ src/components/feature/breadcrumb/index.ts
@@ -1,0 +1,1 @@
+export { ActivityBreadcrumb } from './Activity'
