+++ src/components/feature/index.ts
@@ -1,2 +1,5 @@
 export * from './member-table'
-export { SectionWithTitle } from './SectionWithTitle'
+export * from './breadcrumb'
+export * from './skeleton'
+export * from './section-with-title'
+export * from './semester-pagination'
