+++ src/components/feature/section-with-title/SectionWithTitle.tsx
