+++ src/components/feature/section-with-title/index.ts
@@ -1,0 +1,1 @@
+export { SectionWithTitle } from './SectionWithTitle'
