+++ src/components/feature/semester-pagination/index.ts
@@ -1,0 +1,1 @@
+export { SemesterPagination } from './SemesterPagination'
