+++ src/components/feature/skeleton/Board.tsx
