+++ src/components/feature/skeleton/index.ts
@@ -1,0 +1,3 @@
+export { ActivitySkeleton } from './Activity'
+export { AdminSemesterSkeleton, ActivitySemesterSkeleton } from './Semester'
+export { BoardSkeleton } from './Board'
